refactor: fold remaining session-product row helpers into storage products
Merge the remaining lifecycle, row, and profile helper leaves into the consolidated session-product modules so rebuild and refresh logic reads directly in terms of profiles, timelines, and threads.
