--- polylogue/storage/session_product_lifecycle.py
@@ -1,43 +1,0 @@
-"""Stable lifecycle surface for durable session-product read models."""
-
-from __future__ import annotations
-
-from polylogue.storage.session_product_rebuild import (
-    rebuild_session_products_async,
-    rebuild_session_products_sync,
-)
-from polylogue.storage.session_product_refresh_deletes import (
-    delete_session_products_for_conversation_async,
-    refresh_thread_after_conversation_delete_async,
-)
-from polylogue.storage.session_product_refresh_updates import (
-    refresh_session_products_for_conversation_async,
-)
-from polylogue.storage.session_product_status import (
-    session_product_status_async,
-    session_product_status_sync,
-    session_profile_repair_candidate_ids_async,
-    session_profile_repair_candidate_ids_sync,
-)
-from polylogue.storage.session_product_threads import (
-    thread_conversation_ids_async,
-    thread_conversation_ids_sync,
-    thread_root_id_async,
-    thread_root_id_sync,
-)
-
-__all__ = [
-    "delete_session_products_for_conversation_async",
-    "rebuild_session_products_async",
-    "rebuild_session_products_sync",
-    "refresh_session_products_for_conversation_async",
-    "refresh_thread_after_conversation_delete_async",
-    "session_product_status_async",
-    "session_product_status_sync",
-    "session_profile_repair_candidate_ids_async",
-    "session_profile_repair_candidate_ids_sync",
-    "thread_conversation_ids_async",
-    "thread_conversation_ids_sync",
-    "thread_root_id_async",
-    "thread_root_id_sync",
-]
